+++ bloom/ws/tests/handshake_logging.rs
@@ -1,5 +1,7 @@
 #[path = "common.rs"]
 mod common;
+#[path = "logging_common.rs"]
+mod logging_common;
 
 use bloom_api::ServerToClient;
 use bloom_core::{CreateRoomResult, ParticipantId, RoomId};
@@ -9,12 +11,32 @@ use tokio_tungstenite::connect_async;
 use tokio_tungstenite::tungstenite::protocol::Message;
 
 use common::*;
+use logging_common::*;
+
+fn shared_core_with_arc() -> (
+    SharedCore<MockCore>,
+    std::sync::Arc<std::sync::Mutex<MockCore>>,
+) {
+    let mock_core = MockCore::new(CreateRoomResult {
+        room_id: RoomId::new(),
+        self_id: ParticipantId::new(),
+        participants: vec![],
+    });
+    let core_arc = std::sync::Arc::new(std::sync::Mutex::new(mock_core));
+    let shared = SharedCore::from_arc(core_arc.clone());
+    (shared, core_arc)
+}
 
 /// WSハンドシェイクがHTTP 101で確立され、participant_id付きのspanが出ることを検証する。
 #[tokio::test]
 async fn handshake_returns_switching_protocols_and_sets_participant_span() {
     let (layer, _guard) = setup_tracing();
-    let (server_url, handle) = spawn_bloom_ws_server().await;
+    let core = SharedCore::new(MockCore::new(CreateRoomResult {
+        room_id: RoomId::new(),
+        self_id: ParticipantId::new(),
+        participants: vec![],
+    }));
+    let (server_url, handle) = spawn_bloom_ws_server_with_core(core).await;
 
     let (_ws_stream, response) = connect_async(&server_url)
         .await
@@ -24,7 +46,7 @@ async fn handshake_returns_switching_protocols_and_sets_participant_span() {
 
     tokio::time::sleep(std::time::Duration::from_millis(50)).await;
     let spans = layer.spans.lock().expect("collect spans");
-    assert!(spans_have_field(&spans, "participant_id"));
+    assert!(spans_have_field_value(&spans, "participant_id", ""));
 
     handle.shutdown().await;
 }
@@ -32,8 +54,8 @@ async fn handshake_returns_switching_protocols_and_sets_participant_span() {
 /// CreateRoomを送信するとRoomCreatedが返り、coreが一度呼ばれることを検証。
 #[tokio::test]
 async fn create_room_returns_room_created_and_calls_core_once() {
-    let (layer, _guard) = setup_tracing();
-    let (shared_core, core_arc) = new_shared_core_with_arc();
+    let (_layer, _guard) = setup_tracing();
+    let (shared_core, core_arc) = shared_core_with_arc();
     let (server_url, handle) = spawn_bloom_ws_server_with_core(shared_core).await;
 
     let (mut ws_stream, _response) = connect_async(&server_url)
@@ -53,11 +75,6 @@ async fn create_room_returns_room_created_and_calls_core_once() {
 
     let core_calls = core_arc.lock().expect("lock core").create_room_calls.len();
     assert_eq!(core_calls, 1);
-
-    // span に participant_id が記録されているか軽く確認
-    tokio::time::sleep(std::time::Duration::from_millis(20)).await;
-    let spans = layer.spans.lock().expect("collect spans");
-    assert!(spans_have_field(&spans, "participant_id"));
 
     handle.shutdown().await;
 }
